bullet shield interaction
Added in a spark effect from Bramley Williams's FPS tutorial. Set bitmasks on the shield and projectiles so they interact with each other. Added bullets to shield experiment to test out the effect. So far so good. Need to test orientation of the sparks and maybe code in some response from the shader.