shield bitmasks
Reredid the shield bitmask and bullet bitmask on index 4 so that bullets detect shields, but not vice versa. Moved spark code over to the bullet instead of in the shield script. I also attempted to make the shield shader respond when hit, but that didn't work. It may be that visual shaders aren't meant to work with code and I actually need to write code for the shader.