Shield Component Works!
Stuck shield on an npc.
Added queue_free on death to the shield script.
Had to reattach the healthcomponent to the shield hitbox after saving shield branch as scene.
If the shield is too small, bullets will occasionally pass through and damage the hull, but this is an inaccuracy of the Area3D bullets and not a problem with the shield. Making the shield bigger fixes the issue.