Shield improvements and tests
I tested out that the projectile.gd ignore list actually works and was surprised to learn that it doesn't, because it only triggers on area entered. So I don't need it at all. I can just ignore areas that are entered immediately! I added a check to verify that those areas are attached to shields using a group. I also verified that enemies could shoot the player from within the player's shield without first taking down the shield.